requirements.txt
vosk==0.3.45
pyaudio==0.2.14
pyttsx3==2.90
sentence-transformers==2.2.2
scikit-learn==1.3.0
wikipedia==1.4.0
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2

requests==2.31.0
python-dotenv==1.0.0